requirements.txt
azure-cognitiveservices-speech
pydub